feat: Add `/entry` `/analytics`
-  Add new endpoints to the repo
- Migrate db to use Quarkus FlyAway
- Add analytics and scheduled job
- Use built-in features of metric tracking

+++ 2026/software-engineering/java/quarkus-wellness-api/src/main/java/org/fractalschema/dto/request/CreateEntryRequest.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -18,7 +17,7 @@ public class CreateEntryRequest {
     private LocalDate entryDate = LocalDate.now();
 
     @DecimalMin("0.0") @DecimalMax("24.0")
-    private BigDecimal sleepHours;
+    private Double sleepHours;
 
     @Min(1) @Max(5)
     private Integer sleepQuality;
