+++ 2026/software-engineering/java/quarkus-wellness-api/src/main/java/org/fractalschema/dto/request/UpdateGoalRequest.java
@@ -7,13 +7,12 @@ import lombok.Setter;
 import org.fractalschema.enums.GoalFrequency;
 import org.fractalschema.enums.GoalType;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter @Setter @NoArgsConstructor
 public class UpdateGoalRequest {
     private GoalType goalType;
-    private BigDecimal targetValue;
+    private Double targetValue;
     private GoalFrequency frequency;
     private LocalDate startDate;
     private LocalDate endDate;
